backend/requirements.txt
# Core Dependencies
fastapi
uvicorn
python-dotenv
httpx
email-validator

# Pydantic
pydantic
pydantic-settings
python-multipart

python-jose[cryptography]
passlib[bcrypt]

# Database
psycopg2-binary
sqlalchemy
alembic

# Document Processing
python-docx
pypdf2
docx2txt

redis
celery

//google-generativeai
transformers

# Vector Storage & RAG
chromadb
langchain
langchain-google-genai

# Testing
pytest
pytest-asyncio

#more dependencies
motor
sentence-transformers
langchain-huggingface
langchain-chroma
google-genai
